# perf(cli): perform async temp cleanup