**What's the problem this PR addresses?**

Yarn is cleaning up its temp folders in the `exit` event which forces us to use sync IO calls which has a performance impact.

**How did you fix it?**

Implemented `rmtempPromise` and call it right after the CLI is done and about to exit, where we can run async IO without problems.

**Result**

Time spent cleaning up temp folders after a install with a hot cache and lockfile (`install-cache-and-lock` in the benchmarks)

| | before | after |
|--|--|--|
| Nextjs benchmark | 326ms | 25ms |
| Gatsby benchmark | 654ms | 20ms |
| Typescript Website | 846ms | 44ms |

Benchmark workflow result:
Before: https://github.com/yarnpkg/berry/actions/runs/200276519
After: https://github.com/yarnpkg/berry/actions/runs/200408519

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.